+++ include/sensor/pif_sensor_switch.h
@@ -111,7 +111,7 @@ uint16_t pifSensorSwitch_ProcessAcquire(PifSensorSwitch* p_owner);
  * @param start 즉시 시작할지를 지정한다.
  * @return Task 구조체 포인터를 반환한다.
  */
-PifTask *pifSensorSwitch_AttachTaskAcquire(PifSensorSwitch* p_owner, PifTaskMode mode, uint16_t period, BOOL start);
+PifTask *pifSensorSwitch_AttachTaskAcquire(PifSensorSwitch* p_owner, PifTaskMode mode, uint32_t period, BOOL start);
 
 
 #ifdef PIF_COLLECT_SIGNAL
