+++ source/sensor/pif_sensor_digital.c
@@ -13,7 +13,7 @@
 #endif
 
 
-static uint16_t _doTaskAcquire(PifTask* p_task)
+static uint32_t _doTaskAcquire(PifTask* p_task)
 {
 	pifSensorDigital_ProcessAcquire((PifSensorDigital*)p_task->_p_client);
 	return 0;
@@ -152,7 +152,7 @@ uint16_t pifSensorDigital_ProcessAcquire(PifSensorDigital* p_owner)
 	return 0;
 }
 
-PifTask* pifSensorDigital_AttachTaskAcquire(PifSensorDigital* p_owner, PifTaskMode mode, uint16_t period, BOOL start)
+PifTask* pifSensorDigital_AttachTaskAcquire(PifSensorDigital* p_owner, PifTaskMode mode, uint32_t period, BOOL start)
 {
 	PifTask* p_task;
